fix: avoid indexing extensionless files
