+++ src/core/indexer.rs
@@ -236,97 +236,7 @@ impl Indexer {
     }
 
     fn should_index(&self, path: &Path) -> bool {
-        if let Ok(metadata) = fs::metadata(path) {
-            if metadata.len() > self.max_file_size {
-                return false;
-            }
-        }
-
-        let ext = path
-            .extension()
-            .and_then(|e| e.to_str())
-            .unwrap_or("")
-            .to_lowercase();
-
-        matches!(
-            ext.as_str(),
-            "rs" | "py"
-                | "js"
-                | "ts"
-                | "tsx"
-                | "jsx"
-                | "go"
-                | "c"
-                | "cpp"
-                | "h"
-                | "hpp"
-                | "java"
-                | "kt"
-                | "swift"
-                | "rb"
-                | "php"
-                | "cs"
-                | "fs"
-                | "scala"
-                | "clj"
-                | "ex"
-                | "exs"
-                | "erl"
-                | "hs"
-                | "ml"
-                | "lua"
-                | "r"
-                | "jl"
-                | "dart"
-                | "vue"
-                | "svelte"
-                | "astro"
-                | "html"
-                | "htm"
-                | "css"
-                | "scss"
-                | "sass"
-                | "less"
-                | "json"
-                | "yaml"
-                | "yml"
-                | "toml"
-                | "xml"
-                | "md"
-                | "markdown"
-                | "txt"
-                | "rst"
-                | "tex"
-                | "sh"
-                | "bash"
-                | "zsh"
-                | "fish"
-                | "ps1"
-                | "bat"
-                | "cmd"
-                | "sql"
-                | "graphql"
-                | "proto"
-                | ""
-        ) || path.file_name().is_some_and(|n| {
-            let name = n.to_string_lossy().to_lowercase();
-            matches!(
-                name.as_str(),
-                "dockerfile"
-                    | "makefile"
-                    | "cmakelists.txt"
-                    | "rakefile"
-                    | "gemfile"
-                    | "podfile"
-                    | "vagrantfile"
-                    | ".gitignore"
-                    | ".dockerignore"
-                    | ".env.example"
-                    | "readme"
-                    | "license"
-                    | "changelog"
-            )
-        })
+        super::index_filter::should_index_path(path, self.max_file_size)
     }
 
     fn chunk_content(&self, content: &str) -> Vec<FileChunk> {
@@ -612,97 +522,7 @@ impl ServerIndexer {
     }
 
     fn should_index(&self, path: &Path) -> bool {
-        if let Ok(metadata) = fs::metadata(path) {
-            if metadata.len() > self.max_file_size {
-                return false;
-            }
-        }
-
-        let ext = path
-            .extension()
-            .and_then(|e| e.to_str())
-            .unwrap_or("")
-            .to_lowercase();
-
-        matches!(
-            ext.as_str(),
-            "rs" | "py"
-                | "js"
-                | "ts"
-                | "tsx"
-                | "jsx"
-                | "go"
-                | "c"
-                | "cpp"
-                | "h"
-                | "hpp"
-                | "java"
-                | "kt"
-                | "swift"
-                | "rb"
-                | "php"
-                | "cs"
-                | "fs"
-                | "scala"
-                | "clj"
-                | "ex"
-                | "exs"
-                | "erl"
-                | "hs"
-                | "ml"
-                | "lua"
-                | "r"
-                | "jl"
-                | "dart"
-                | "vue"
-                | "svelte"
-                | "astro"
-                | "html"
-                | "htm"
-                | "css"
-                | "scss"
-                | "sass"
-                | "less"
-                | "json"
-                | "yaml"
-                | "yml"
-                | "toml"
-                | "xml"
-                | "md"
-                | "markdown"
-                | "txt"
-                | "rst"
-                | "tex"
-                | "sh"
-                | "bash"
-                | "zsh"
-                | "fish"
-                | "ps1"
-                | "bat"
-                | "cmd"
-                | "sql"
-                | "graphql"
-                | "proto"
-                | ""
-        ) || path.file_name().is_some_and(|n| {
-            let name = n.to_string_lossy().to_lowercase();
-            matches!(
-                name.as_str(),
-                "dockerfile"
-                    | "makefile"
-                    | "cmakelists.txt"
-                    | "rakefile"
-                    | "gemfile"
-                    | "podfile"
-                    | "vagrantfile"
-                    | ".gitignore"
-                    | ".dockerignore"
-                    | ".env.example"
-                    | "readme"
-                    | "license"
-                    | "changelog"
-            )
-        })
+        super::index_filter::should_index_path(path, self.max_file_size)
     }
 
     fn chunk_content(&self, content: &str) -> Vec<FileChunk> {
